+++ patch-jdeploy-before.js
@@ -15,26 +15,24 @@ else {
     fse.mkdirSync('dist');
 }
 
-// Copy files into dist folder
-fse.copyFileSync('src/index.js', 'dist/index.js');
-console.info('NGL: Copied src/index.js into dist folder');
+// Copy files into dist folder (where they will be taken by "jdeploy install" command and copied to jdeploy_bundle)
+const filesToCopy =
+    [
+        "index.js",
+        "groovy-lint.js",
+        "groovy-lint-fix.js",
+        "groovy-lint-rules.js",
+        "options.js",
+        "utils.js"
+    ];
 
-fse.copyFileSync('src/groovy-lint.js', 'dist/groovy-lint.js');
-console.info('NGL: Copied groovy-lint.js.js into dist folder');
-
-fse.copyFileSync('src/options.js', 'dist/options.js');
-console.info('NGL: Copied options.js into dist folder');
-
-fse.copyFileSync('src/groovy-lint-fix.js', 'dist/groovy-lint-fix.js');
-console.info('NGL: Copied groovy-lint-fix.js into dist folder');
-
-fse.copyFileSync('src/groovy-lint-rules.js', 'dist/groovy-lint-rules.js');
-console.info('NGL: Copied src/groovy-lint-rules.js into dist folder');
+for (const fileName of filesToCopy) {
+    fse.copyFileSync('src/' + fileName, 'dist/' + fileName);
+    console.info('NGL: Copied lib files into dist/lib');
+}
 
 fse.copySync('lib', 'dist/lib');
 console.info('NGL: Copied lib files into dist/lib');
-
-fse.copy
 
 process.exit(0);